fix the following warnings, introduced in r6571 (ticket #401): {{{ fs.c: In function 'gfm_server_replica_lost': fs.c:4072: warning: format '%lld' expects type 'long long int', but argument 7 has type 'gfarm_ino_t' fs.c:4072: warning: format '%lld' expects type 'long long int', but argument 8 has type 'gfarm_uint64_t' fs.c: In function 'gfm_server_replica_create_file_in_lost_found': fs.c:4269: warning: format '%lld' expects type 'long long int', but argument 7 has type 'gfarm_ino_t' fs.c:4269: warning: format '%lld' expects type 'long long int', but argument 8 has type 'gfarm_uint64_t' fs.c:4269: warning: format '%lld' expects type 'long long int', but argument 10 has type 'gfarm_ino_t' fs.c:4269: warning: format '%lld' expects type 'long long int', but argument 11 has type 'gfarm_uint64_t' }}}
pointed out by tatebe@-san

git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@6577 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfmd/fs.c
@@ -5090,7 +5090,8 @@ gfm_server_replica_lost(struct peer *peer, gfp_xdr_xid_t xid, size_t *sizep,
 		if (e == GFARM_ERR_NO_ERROR) {
 			gflog_notice(GFARM_MSG_UNFIXED,
 			    "inode %lld:%lld on %s: invalid metadata deleted",
-			    inum, gen, host_name(spool_host));
+			    (long long)inum, (long long)gen,
+			    host_name(spool_host));
 		}
 	}
 
@@ -5315,8 +5316,9 @@ gfm_server_replica_create_file_in_lost_found(struct peer *peer,
 			gflog_notice(GFARM_MSG_UNFIXED,
 			    "inode %lld:%lld on %s -> %lld:%lld: "
 			    "moved to lost+found",
-			    inum_old, gen_old, host_name(spool_host),
-			    inum_new, gen_new);
+			    (long long)inum_old, (long long)gen_old,
+			    host_name(spool_host),
+			    (long long)inum_new, (long long)gen_new);
 		}
 	} else
 		e_rpc = GFARM_ERR_NO_ERROR;
